Consolidate routing logic
